+++ Cider Remote/Data/Network.swift
@@ -1,0 +1,10 @@
+// Made by Lumaa
+
+import Foundation
+
+enum NetworkError: Error {
+    case invalidURL
+    case invalidResponse
+    case decodingError
+    case serverError(String)
+}
