+++ Cider Remote/Data/Track.swift
@@ -1,0 +1,12 @@
+// Made by Lumaa
+
+import Foundation
+
+struct Track: Codable, Equatable {
+    let id: String
+    let title: String
+    let artist: String
+    let album: String
+    let artwork: String
+    let duration: Double
+}
