+++ Cider Remote/Views/MusicPlayerView.swift
@@ -163,23 +163,9 @@ struct MusicPlayerView: View {
 
 extension MusicPlayerViewModel {
     func initializePlayer() async {
-        do {
-            try await getCurrentTrack()
-            try await getCurrentVolume()
-            try await fetchQueueItems()
-        } catch {
-            await MainActor.run {
-                self.handleError(error)
-            }
-        }
-    }
-}
-
-struct ScaleButtonStyle: ButtonStyle {
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .scaleEffect(configuration.isPressed ? 0.9 : 1)
-            .animation(.easeInOut(duration: 0.2), value: configuration.isPressed)
+        await getCurrentTrack()
+        await getCurrentVolume()
+        await fetchQueueItems()
     }
 }
 
@@ -714,15 +700,6 @@ struct PlayerControlsView: View {
     }
 }
 
-struct ResponsiveButtonStyle: ButtonStyle {
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .scaleEffect(configuration.isPressed ? 0.9 : 1.0)
-            .opacity(configuration.isPressed ? 0.6 : 1.0)
-            .animation(.easeInOut(duration: 0.1), value: configuration.isPressed)
-    }
-}
-
 struct VolumeControlView: View {
     @ObservedObject var viewModel: MusicPlayerViewModel
     @State private var isDragging = false
@@ -746,128 +723,6 @@ struct VolumeControlView: View {
                 .foregroundColor(.secondary)
         }
         .frame(height: 30)  // Set a fixed height for the volume control
-    }
-}
-
-
-struct CustomSlider: View {
-    @Binding var value: Double
-    @EnvironmentObject var colorScheme: ColorSchemeManager
-    let bounds: ClosedRange<Double>
-    @Binding var isDragging: Bool
-    let onEditingChanged: (Bool) -> Void
-
-    @State private var lastDragValue: Double?
-
-    var body: some View {
-        GeometryReader { geometry in
-            ZStack(alignment: .leading) {
-                Rectangle()
-                    .fill(Material.ultraThin)
-                    .frame(height: 8)
-
-                Rectangle()
-                    .fill(Color.white.opacity(0.5))
-                    .frame(width: CGFloat((value - bounds.lowerBound) / (bounds.upperBound - bounds.lowerBound)) * geometry.size.width, height: 8)
-            }
-            .clipShape(Capsule())
-            .frame(height: geometry.size.height)
-            .contentShape(Rectangle())
-            .gesture(
-                DragGesture(minimumDistance: 0)
-                    .onChanged { gestureValue in
-                        isDragging = true
-                        let newValue = bounds.lowerBound + (bounds.upperBound - bounds.lowerBound) * Double(gestureValue.location.x / geometry.size.width)
-                        value = max(bounds.lowerBound, min(bounds.upperBound, newValue))
-
-                        // Haptic feedback
-                        if let last = lastDragValue, abs(newValue - last) > (bounds.upperBound - bounds.lowerBound) / 100 {
-                            let impact = UIImpactFeedbackGenerator(style: .light) //MARK: API is deprecated
-                            impact.impactOccurred()
-                            lastDragValue = newValue
-                        } else if lastDragValue == nil {
-                            lastDragValue = newValue
-                        }
-                    }
-                    .onEnded { _ in
-                        isDragging = false
-                        lastDragValue = nil
-                        onEditingChanged(false)
-                    }
-            )
-        }
-        .environment(\.colorScheme, ColorScheme.light)
-        .frame(height: 44)
-    }
-}
-
-struct LargeButton: View {
-    let title: String
-    let systemImage: String
-    let action: () -> Void
-    let size: ElementSize
-    let geometry: GeometryProxy
-
-    var body: some View {
-        Button(action: action) {
-            HStack {
-                Image(systemName: systemImage)
-                Text(title)
-            }
-            .font(.system(size: adjustedFontSize))
-            .foregroundColor(.primary)
-            .frame(maxWidth: .infinity)
-            .frame(height: adjustedHeight)
-            .background(Color.secondary.opacity(0.2))
-            .cornerRadius(8)
-        }
-    }
-
-    private var adjustedFontSize: CGFloat {
-        min(size.fontSize * 0.8, 22)  // Reduce font size and set a maximum
-    }
-
-    private var adjustedHeight: CGFloat {
-        min(size.dimension * 1.2, 60)  // Adjust height and set a maximum
-    }
-}
-
-struct SmallButton: View {
-    let title: String
-    let systemImage: String
-    let action: () -> Void
-    let size: ElementSize
-    let geometry: GeometryProxy
-
-    var body: some View {
-        Button(action: action) {
-            HStack {
-                Image(systemName: systemImage)
-                Text(title)
-            }
-            .font(.system(size: adjustedFontSize))
-            .foregroundColor(.primary)
-            .frame(maxWidth: .infinity)
-            .frame(height: adjustedHeight)
-            .background(Color.secondary.opacity(0.2))
-            .cornerRadius(8)
-        }
-    }
-
-    private var adjustedFontSize: CGFloat {
-        switch size {
-        case .small: return 12
-        case .medium: return 14
-        case .large: return 16
-        }
-    }
-
-    private var adjustedHeight: CGFloat {
-        switch size {
-        case .small: return 30
-        case .medium: return 34
-        case .large: return 38
-        }
     }
 }
 
@@ -1027,32 +882,6 @@ struct BlurredImageView: View {
             .frame(maxWidth: UIScreen.main.bounds.width, maxHeight: UIScreen.main.bounds.height)
             .blur(radius: 60)
     }
-}
-
-struct Track: Codable, Equatable {
-    let id: String
-    let title: String
-    let artist: String
-    let album: String
-    let artwork: String
-    let duration: Double
-}
-
-struct SpringyButtonStyle: ButtonStyle {
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .contentShape(Rectangle()) // Makes the entire frame tappable
-            .scaleEffect(configuration.isPressed ? 0.9 : 1.0)
-            .opacity(configuration.isPressed ? 0.6 : 1.0)
-            .animation(.spring(response: 0.3, dampingFraction: 0.6, blendDuration: 0), value: configuration.isPressed)
-    }
-}
-
-enum NetworkError: Error {
-    case invalidURL
-    case invalidResponse
-    case decodingError
-    case serverError(String)
 }
 
 @MainActor
